requirements.txt
numpy~=1.23.5
pillow~=9.4.0
tensorflow~=2.7.0
matplotlib~=3.6.2
scikit-learn~=1.1.3
pandas~=1.5.3
requests~=2.30.0